New version with functional checkpoint and restore, complete log in chronological order below. Final patch by far the largest and most important, since it heavily edits memckpt.c
Summary:
Cleanup patch for the userspace tools: Added error messages to notify when they
are being misused, and also merged the data structures being passed to the kernel.
This in turn lets us factor out code common to both ioctl() calls.

removed '-1 is this process' path

In slsmm.c passing -1 to the PID argument means 'this process'. This can be done from userspace with a getpid() call

change locking scheme in slsmm.c

Change the locking scheme in the main module to be more straightforward.

added comments to slsmm.c

Added comments to the main module regarding future work and possible issues.

implement restore functionality

Restore now works.

Test Plan: The loop.c program in the tests/ directory has the expected behavior when checkpointed and restored.

Reviewers: ali, hongbo

Differential Revision: http://review.rcs.uwaterloo.ca/D15