Single Level Object Store (SLOS)
Summary:
The Single Level Object Store (SLOS) backing the Single Level Store (SLS) part of Aurora. The SLOS is an extent-based, record oriented filesytem which supports CoW semantics.

The SLOS object store holds a collection of inodes, identified by a unique key. This key normally represents the PID of the process being checkpointed, but can be repurposed to signify
arbitrary kinds of values. Each inode holds a collection of records, which look to inodes like contiguous areas of data with specific types. The types are pieces of metadata used by
the users of SLOS to specify what each kind of data each type of value represents.

Records implemented as ordered collections of extents on the disk. They are implemented using btrees to hold the extents allocated for the records. Records can be sparse, with holes
being read as zeros. Reads and writes are done with byte granularity. The semantic gap between the extents' block-granularity semantics and that of the records' is bridged by specifying
exactly which bytes in an extent are actually in use by the record.

The bnodes are primarily on-disk structures. They are read from and written to disk using the slos_{read, write}() subroutines. These are equivalent to the operations done by other filesystems
when flushing to and from the disk, like ffs_read(), simplified for now.

Inodes and records are allocated by using an allocator based on a btree. This allocator is comprised of two btrees. One btree uses as keys the offsets of all free extents, and holds
as values their sizes. The other is keyes by extent size, and has as a value pointers to buckets of free extents with the given size. The buckets are themselves btrees.

Btrees are an abstraction on top of bnodes, specifying the usual operations like insert, delete, and search, together with balancing operations such as node splits and merges. Bnodes
are block-size data structures with variable-length fields. Each bnode starts with metadata that is used by each operation to adapt to the bnode's individual layout, followed by
an array of 64bit keys. Depending on whether a bnode is internal or external, it then holds either an array of disk pointers to its child bnodes, or an array of values of sizes specified
at node creation time. The size of these arrays is as large as possible for a node to hold an equal amount of keys and values, or disk pointers, whichever is greater.

Test Plan: Each component has a different test, typically a generator of random operations using the component's API. For example, the btree component is testing by randomly inserting, searching, deleting, and so on. This approach has proved effective, to the point that each component can be used for arbitrary amounts of time without misbehaving. These tests are not used in  normal operation, but can be turned on if bugs are found.

Reviewers: ali, ryan

Reviewed By: ali, ryan

Subscribers: jenkins

Differential Revision: https://review.rcs.uwaterloo.ca/D147