seeking and iterating
Summary:
In order for the SLOS to be able to used by the SLS, it must
be possible to seek the non-empty parts of records. Moreover, inodes should
be able to be traversed using a standard SLOS-provided API. This patch provides
the functionality necesary for these operations; both of them are implemented using
simple logic checks in conjunction with standard btree operations.

Test Plan:
Added testing for the seek operation in slos_record.c. Only seeking
holes can be checked, due to the way the test works.

Reviewers: ali, ryan, hongbo

Reviewed By: ali

Subscribers: jenkins

Differential Revision: https://review.rcs.uwaterloo.ca/D156