new bottom layer for SLS
Summary:
The old bottom layer for the SLS was built with files in mind. SLOS support was built on top of the existing code due to time constraints, causing the code to be brittle. Moreover, the first version of SLS assumed single processes being checkpointed. This assumption had important ramifications for our design decisions, e.g. the way objects and data were being checkpointed.

A first step towards the new version is this alternative lower layer for the SLS. It is not yet hooked up to the rest of the code, but it does have its own tests, and it uses the SLOS properly.

Test Plan: Builtin test

Reviewers: ali, ryan

Reviewed By: ali

Subscribers: jenkins

Differential Revision: https://review.rcs.uwaterloo.ca/D190