changed memory checkpointing method
Summary: The previous version of SLS checkpointed memory in terms of address spaces. This does not work with multiple processes, since  we are unable to discern what data is shared by just using the VM entries. Change the code so that memory checkpointing is done in terms of VM Objects, from which we track the processes that are currently using them.

Test Plan: Using the SLS

Reviewers: ali, ryan

Reviewed By: ryan

Subscribers: jenkins

Differential Revision: https://review.rcs.uwaterloo.ca/D198