multiprocess checkpointing
Summary: The SLS was originally built to handle only single processes at a time. That meant that groups of processes that shared anonymous resources such as memory and pipes could not be restored properly, because there was no way of discerning which resources were being shared and how. Now that memory and pipes are deduplicated and properly shared at restore time (D201), modify the checkpoint and restore paths so that they allow for multiple processes.

Test Plan: Same ol' tests, plus sharemap/pipe

Reviewers: ali, ryan

Reviewed By: ali

Differential Revision: https://review.rcs.uwaterloo.ca/D204