mosh-server fixes
Summary:
One of our benchmarks is mosh, a remote shell which uses UDP to communicate between the client and the server. Add enough functionality to allow for its checkpoint and restore. Functionality includes:
- Allowing for UDP listening sockets to be checkpointed and restored
- Checkpoint flag for including all of the SLS processes' descendants in the checkpoint
- Restore flag that restores the process as a daemon
- Restore option for restoring the processes as children of the original slsctl restore (used to attach them to the latter's terminal and for allowing for sending them signals like Ctrl+C)

Test Plan: Running the tests and verifying that they work properly, also using mosh interactively and verifying it works. NOTE: Sometimes the pts used between mosh and the child is not restored correctly; even in that case, the failure is graceful.

Reviewers: ali, ryan

Reviewed By: ali

Differential Revision: https://review.rcs.uwaterloo.ca/D212