began optimizations
Summary:
Began implementing the necessary optimizations to reduce the checkpointing overhead. The bottlenecks here are two: Blowing away the mappings is costlyin terms of stop time, and blowing away the mappings causes enough minor page faults to cause significant performance degradation. Therefore, we use a custom page protection function that exists in our modified kernel, pmap_protect_pglist(), and try to repair the mappings of the process by eagerly copying it from Aurora while the process itself is executing.

Note that the optimizations right now DO NOT DO MUCH. In fact, eagerly copying the pages online is _actively_harmful_, because it locks up the VM objects, and causes the process to wait for Aurora before it services its own page faults. Still, this initial code will be used as a springboard for implementing an heuristic with which we selectively repair mappings based on FreeBSD's CLOCK algorithm.

Test Plan: Same old

Reviewers: ryan, tavianator, ali

Reviewed By: ryan

Differential Revision: https://review.rcs.uwaterloo.ca/D254