Fix IO redirection, timestamps and a few potential bugs.
Summary:
IO redirection in the console causes applications to use several Ioctls to test if the file descriptor is a tty or not.  In these cases we should return ENOTTY, otherwise we return ENOSYS for FIOSEEKDATA/HOLE until those are implemented.

Bonus fixed the inode timestamps and a few potential bugs in the code.

Test Plan: manual testing

Reviewers: ryan, etsal, tavianator

Reviewed By: tavianator

Differential Revision: https://review.rcs.uwaterloo.ca/D276