Fix memory leaks
Summary: This change fixes the three memory leaks remaining in the system.  Two slos node objects for allocation and the root node.  Bonus I found an unreported leak of the root vnode and added the vrele for that as well.  Seems that vrele for VV_SYSTEM nodes does not call reclaim.

Test Plan: test scripts

Reviewers: ryan, etsal, tavianator

Reviewed By: etsal

Differential Revision: https://review.rcs.uwaterloo.ca/D279