bring SLS table tests up to date
Summary: In order to be able to handle large workloads we need to be certain the data received from disk is valid. Bring the tests in sls_table.c up to date with the rest of the SLS in order to automatically be able to test the disk IO functions from the SLS side.

Test Plan: Running the tests in sls_test.c at load time by using the -DTEST flag when compiling

Reviewers: ryan, tavianator, ali

Reviewed By: ryan

Differential Revision: https://review.rcs.uwaterloo.ca/D284