invalid psind bugfix
Summary: There is a bug where using the psind as an index into the pagesize array leads to page sizes of 0, which translates into what amounts to a memory corruption bug. Use a default value for the size of the page when pagesizes[psind] is 0.

Test Plan: Running memtester and llist

Reviewers: ali, ryan, tavianator

Reviewed By: ali

Differential Revision: https://review.rcs.uwaterloo.ca/D308