do not truncate files to 0 when rendered inactive
Summary: Files that have no hard links to them in the filesystem are considered inactive and inaccessible by filesystems, and are thus destroyed. However, in the SLOS we might still be able to access them by a checkpointed process. Do not call truncate() for inactive files in the SLOS.

Test Plan: Unlink a file, check with fsdb that it's still there

Reviewers: ryan, ali, tavianator

Reviewed By: tavianator

Differential Revision: https://review.rcs.uwaterloo.ca/D331