restoring Firefox benchmark
Summary:
Add fixes that allow us to restore the Firefox benchmark. The latter is comprised of a Python script that uses Selenium 3 to spawn a geckodriver process, which in turn spawns Marionette Firefox processes and runs the Kraken JS benchmark on them. Add fixes which allow us torestore the whole setup, including:

- Checkpointing kqueues while in the KQ_SEL state; this is fine because no one is selecting onthe kqueue, even if the bit is set.
- Restoring vnodes by SLOS inode instead of path when appropriate. This is used for unlinked vnodes that are still in the SLOS.
- Remove file truncation to 0 when a file is considered inactive in the SLOS. There might be checkpoints pointing to it, so the data is still accessible.

Test Plan: Restoring the Firefox benchmark

Reviewers: tavianator, ali, ryan

Reviewed By: ali

Differential Revision: https://review.rcs.uwaterloo.ca/D338