fixes for nginx
Summary:
There are some paths in the code that are exercised by nginx, but not by our other benchmarks:
- The master process and the nginx workers are in the same process group
- There are detached knotes present in the kqueues at checkpoint time

Also make sure we cannot add the kernel process to SLS partitions. Trying to checkpoint the kernel rightfully crashes, so do not allow attempts to do that.

Test Plan: Checkpointing and restoring nginx

Reviewers: ali, tavianator, ryan

Reviewed By: ali

Differential Revision: https://review.rcs.uwaterloo.ca/D340