scripting functions and bugfixes
Summary: We need to automatically do multiple runs for each benchmark, so that we can take numbers. We have also introduced a bug in D360 that stops the SLS from checkpointing to disk more than once for each run, since calls to slos_new_node() fail when asking for a specific inode that already exists. Add the Python code necessary for benchmarking, and fix the bug.

Test Plan: Running the run.py script and finally taking numbers

Reviewers: ali, tavianator, ryan

Reviewed By: ali

Differential Revision: https://review.rcs.uwaterloo.ca/D367