synchronous checkpointing
Summary: Checkpointing is normally initiated externally, that is, the process being checkpointed is not the one making the calls to the SLS. The checkpointing is done by a kernel thread attached to PID 0, so this means the process initiating the checkpoint immediately returns after doing so. This is problematic for a number of cases, including when trying to benchmark checkpointing and dump times by using sls_checkpoint() calls from within the benchmark. Add a new flag to the checkpoint call allowing for synchronous checkpointing. Synchronously checkpointing means stopping the process (partition in the case of >1 processes) when the call is initiated, and only letting it continue after the kernel thread is done. By using different backends (memory, asynchronous dumping to disk, synchronous dumping to disk) we can use this mode to measure the time needed for in-memory checkpointing and initiating the dumping, as well as total dump time.

Test Plan: Using the new vmobject benchmark to measure the latency of the SLS module

Reviewers: ali, tavianator, ryan

Reviewed By: ali

Differential Revision: https://review.rcs.uwaterloo.ca/D369