added the option to ignore unlinked files
Summary: A lot of the benchmarks make use of unlinked files, which currently crash the SLS if it attempts to checkpoint them. Unlinked files need to be moved to the SLOS in order to ensure the proper execution of restored processes, but are not needed for benchmarking. Add the option to ignore unlinked files when checkpointing to speed up the development and benchmarking process

Test Plan: Running workloads like Apache Tomcat, which use unlinked files outside of the SLOS

Reviewers: ali, tavianator, ryan

Reviewed By: ali

Differential Revision: https://review.rcs.uwaterloo.ca/D371