refactored process restore code, ckpt/rest thread spawning
Summary:
Checkpointing is currently being done by a separate thread that signals the one doing the ioctl. Checkpointing can be done asynchronously or synchronously; in the former case the ioctl thread returns immediately after spawning the kernel thread, while in the latter it waits until the checkpointed processes are resumed. Synchronous mode was added to easily measure stop times, but that meant missing errors that happened during IO. Since we can easily use Dtrace to measure stop times, remove synchronous mode as a separate mode, and modify it so that the thread doing the ioctl only returns after the checkpointing thread is fully done.

Restoring was designed to be done asynchronously, but is currently done by the thread doing the ioctl. Modify the code so it's done by a separate thread, so that in the future a thread doing a restore can spawn multiple restore threads restoring multiple partitions at once, and waiting for all of them.

The slsrest_proc() process restore function is currently a monolithic blob that is difficult to follow, even if it is actually a sequence of independent operations. Refactor it by extracting each operation into its own function to make it more understandable.

Test Plan:
Running the testbench:
Build: FreeBSD 12.1-RELEASE FASTDBG
Test started: Sun Nov 15 20:22:15 EST 2020
Test                                       Status      Time
-----------------------------------------------------------
001-load-unload.sh                       [ success ]
002-load-mount-unload.sh                 [ success ]
003-mount-unmount.sh                     [ success ]
004-load-double-sls.sh                   [ success ]
005-unload-double-sls.sh                 [ success ]
006-load-double-slos.sh                  [ success ]
007-unload-double-slos.sh                [ success ]
008-mount-double.sh                      [ success ]
009-unload-slos-before-sls.sh            [ success ]
010-unmount-before-sls.sh                [ success ]
101-simple-file-io.sh                    [ success ]
201-simple-mem-ckpt.sh                   [ success ]
202-simple-osd-ckpt.sh                   [ success ]
203-busy-loop.sh                         [ success ]
204-print-stdout.sh                      [ success ]
205-array-traversal.sh                   [ success ]
206-mmap-file.sh                         [ success ]
207-mmap-anonymous.sh                    [ success ]
208-signal-handler.sh                    [ success ]
209-pthreads-mutex.sh                    [ success ]
210-file-vnode.sh                        [ success ]
211-fork-simple.sh                       [ success ]
212-file-fifo.sh                         [ success ]
213-fork-deltas.sh                       [ success ]
214-file-pipes.sh                        [ success ]
215-file-posixshm.sh                     [ success ]
216-sysv-shm.sh                          [ success ]
217-file-udplisten.sh                    [ success ]
218-file-tcplisten.sh                    [ success ]
219-file-socketpair.sh                   [ success ]
220-fork-shared-mapping.sh               [ success ]
221-lazy-osd-ckpt.sh                     [ success ]
222-self-ckpt.sh                         [ success ]

Reviewers: ali, tavianator, ryan

Reviewed By: ali

Maniphest Tasks: T60

Differential Revision: https://review.rcs.uwaterloo.ca/D390