calling accept() in Metropolis mode now exits after checkpointing
Summary:
Metropolis mode was originally supposed to work with the function calling accept() on connections by the local daemon. With socket handoffs enabled, this is not the case; have the function exit after calling

NOTE: This only works for single process functions; we actually need to go through all processes in the partition and kill them.

Test Plan: Running the testbench without regressions (235 has been modified)

Reviewers: ali, tavianator, ryan

Reviewed By: ryan

Differential Revision: https://review.rcs.uwaterloo.ca/D428