change SRCROOT in testbench to a full path instead of a relative one
Summary: SRCROOT right now has the value "..", which is not expanded to a full path, and thus changes value with every directory change. This behavior breaks the testbench when executing any scripts that change their working directory. Change the call to expand to a full path instead.

Test Plan: Checking that the testbench is still fine

Reviewers: ali, tavianator, ryan

Reviewed By: ali

Differential Revision: https://review.rcs.uwaterloo.ca/D434