fix load/unload and mount/unmount tests to prevent inconsistent state
Summary:
The SLS assumes that it has an underlying file system to which it can send data, but the code currently allows us to partially or completely unmount while the SLS is running.
Fix edge case, and also ensure that the FS prevents the SLOS from unloading or running any teardown routines.

Test Plan: Running the testbench:

Reviewers: ali, tavianator, ryan

Reviewed By: ali

Differential Revision: https://review.rcs.uwaterloo.ca/D621