[3/6] added userspace integrity checking for radix tree based inodes
Summary:
It is difficult to track down file system bugs by debugging
directly in the kernel. The existing debugger parses btrees and has
no knowledge of the new radix tree data structure. Add a userspace
utility for traversing a subset of the file system's metadata as a
form of integrity checking.

Test Plan: using the utility

Reviewers: ali, tavianator, ryan

Reviewed By: ryan

Differential Revision: https://review.rcs.uwaterloo.ca/D697