[5/6] added sparse radix tree traversal to support SLS operations
Summary:
The SLS requires enumerating all extents in a sparse inode
when eagerly restoring application memory. The radix tree does not
implement this functionality because it is unnecessary in the file
system. Add a keymin() operation to the radix tree for fully
traversing a sparse tree, properly skipping all holes.

Test Plan: testbench

Reviewers: ali, tavianator, ryan

Reviewed By: ryan

Differential Revision: https://review.rcs.uwaterloo.ca/D695